+++ src/parser/dissectors/transport/dispatch.ts
@@ -3,8 +3,18 @@ import { parseTCP } from "./tcp/parser";
 import { parseUDP } from "./udp/parser";
 import { parseICMPv4 } from "./ICMPv4/parser";
 import { parseICMPv6 } from "./ICMPv6/parser";
+import type { TCPPacket } from "./tcp/types";
+import type { IPv4Packet } from "../network/ipv4/types";
+import type { IPv6Packet } from "../network/ipv6/types";
 
 type TransportParser = (raw: Uint8Array) => TransportLayer;
+
+
+type TransportHook = (tcp: TCPPacket, ip: IPv4Packet | IPv6Packet) => void;
+let hook: TransportHook | null = null;
+
+export function setTransportHook(fn: TransportHook) { hook = fn };
+
 
 const transportRegistry: Partial<Record<number, TransportParser>> = {
   1: parseICMPv4,
@@ -14,7 +24,7 @@ const transportRegistry: Partial<Record<number, TransportParser>> = {
 }
 
 
-export function dispatchTransport(protocol: number, raw: Uint8Array): TransportLayer {
+export function dispatchTransport(protocol: number, ip: IPv4Packet | IPv6Packet, raw: Uint8Array): TransportLayer {
   const parser = transportRegistry[protocol];
 
   if(!parser) {
@@ -24,5 +34,11 @@ export function dispatchTransport(protocol: number, raw: Uint8Array): TransportL
     } as UnknownTransportLayer
   }
 
-  return parser(raw)
+  const result = parser(raw);
+
+  if(result.type === "tcp" && hook) {
+    hook(result, ip)
+  }
+
+  return result;
 }